LRCI-3892 Remove unnecessary JobQueue component
